requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
bitarray=2.4.0=pypi_0
blas=1.0=mkl
bpemb=0.3.3=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py39h06a4308_2
cffi=1.15.0=pypi_0
charset-normalizer=2.0.11=pypi_0
click=8.0.3=pypi_0
colorama=0.4.4=pypi_0
conllu=4.4.1=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
cython=0.29.28=pypi_0
deprecated=1.2.13=pypi_0
fairseq=1.0.0a0+d03f4e7=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.4.2=pypi_0
flair=0.10=pypi_0
fonttools=4.29.1=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=pypi_0
fsspec=2022.2.0=pypi_0
ftfy=6.1.1=pypi_0
future=0.18.2=pypi_0
gdown=3.12.2=pypi_0
gensim=4.1.2=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
huggingface-hub=0.4.0=pypi_0
hydra-core=1.0.7=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.2=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
janome=0.4.1=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
kiwisolver=1.3.2=pypi_0
konoha=4.6.5=pypi_0
lame=3.100=h7b6447c_0
langdetect=1.0.9=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lxml=4.7.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pypi_0
matplotlib=3.5.1=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
more-itertools=8.8.0=pypi_0
mpld3=0.3=pypi_0
multidict=6.0.2=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyhd3eb1b0_0
omegaconf=2.0.6=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1m=h7f8727e_0
overrides=3.1.0=pypi_0
packaging=21.3=pypi_0
pandas=1.4.0=pypi_0
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
portalocker=2.4.0=pypi_0
protobuf=3.19.4=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pydeprecate=0.3.1=pypi_0
pyparsing=3.0.7=pypi_0
pysocks=1.7.1=pypi_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pypi_0
pytorch=1.10.2=py3.9_cuda10.2_cudnn7.6.5_0
pytorch-lightning=1.5.10=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.1.18=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
sacrebleu=2.0.0=pypi_0
sacremoses=0.0.47=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
segtok=1.5.11=pypi_0
sentencepiece=0.1.95=pypi_0
setuptools=59.5.0=pypi_0
six=1.16.0=pyhd3eb1b0_0
smart-open=5.2.1=pypi_0
sqlite=3.37.2=hc218d9a_0
sqlitedict=1.7.0=pypi_0
tabulate=0.8.9=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.11.4=pypi_0
torchaudio=0.10.2=py39_cu102
torchmetrics=0.7.2=pypi_0
torchvision=0.11.3=py39_cu102
tqdm=4.62.3=pypi_0
transformers=4.16.2=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.8=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=2.0.3=pypi_0
wget=3.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wikipedia-api=0.5.4=pypi_0
wrapt=1.13.3=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
zipp=3.7.0=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
